webSecurityAdded

+++ src/main/java/com/msa/bankingsystem/api/controllers/AccountsController.java
@@ -7,6 +7,7 @@ import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,12 +22,13 @@ import com.msa.bankingsystem.core.mapper.AccountDtoMapper;
 import com.msa.bankingsystem.core.results.DataResult;
 import com.msa.bankingsystem.core.results.Result;
 import com.msa.bankingsystem.models.Account;
-import com.msa.bankingsystem.services.IAccountService;
+import com.msa.bankingsystem.services.account.IAccountService;
 import com.msa.bankingsystem.services.dtos.GetListAccountDto;
 import com.msa.bankingsystem.services.dtos.GetListLogDto;
 import com.msa.bankingsystem.services.requests.CreateAccountRequest;
 import com.msa.bankingsystem.services.requests.CreateDepositRequest;
 import com.msa.bankingsystem.services.requests.CreateTransferRequest;
+import com.msa.bankingsystem.services.securityConfig.MyCustomUser;
 
 @RestController
 @RequestMapping(path = "/api")
@@ -56,13 +58,22 @@ public class AccountsController {
 	private ResponseEntity<DataResult<GetListAccountDto>> getAccountByAccountId(
 			@PathVariable(name = "accountNumber") String accountNumber) {
 
+		DataResult<GetListAccountDto> dataResult;
+
+		if (checkAuthUser(accountNumber)) {
+
+			dataResult = new DataResult<GetListAccountDto>(null, false, "Invalid Account Number");
+			return new ResponseEntity<DataResult<GetListAccountDto>>(dataResult, HttpStatus.FORBIDDEN);
+		}
+
 		DataResult<Account> account = this.iAccountService.getByAccountNumber(accountNumber);
+
 		if (!account.isSuccess() || account == null) {
 			return new ResponseEntity<>((DataResult) account, HttpStatus.NOT_FOUND);
 		}
-		DataResult<GetListAccountDto> dataResult = new DataResult<GetListAccountDto>(
-				this.accountDtoMapper.accountToAccountDto(account.getData()), account.isSuccess(),
-				account.getMessage());
+
+		dataResult = new DataResult<GetListAccountDto>(this.accountDtoMapper.accountToAccountDto(account.getData()),
+				account.isSuccess(), account.getMessage());
 
 		return ResponseEntity.ok().body(dataResult);
 
@@ -73,15 +84,22 @@ public class AccountsController {
 			@PathVariable(name = "accountNumber") String accountNumber,
 			@RequestBody @Valid CreateDepositRequest createDepositRequest) {
 
+		DataResult<GetListAccountDto> dataResult;
+
+		if (checkAuthUser(accountNumber)) {
+
+			dataResult = new DataResult<GetListAccountDto>(null, false, "Invalid Account Number");
+			return new ResponseEntity<DataResult<GetListAccountDto>>(dataResult, HttpStatus.FORBIDDEN);
+		}
+
 		DataResult<Account> account = this.iAccountService.deposit(accountNumber, createDepositRequest);
 
 		if (!account.isSuccess() || account == null) {
 			return new ResponseEntity<>((DataResult) account, HttpStatus.NOT_FOUND);
 		}
 
-		DataResult<GetListAccountDto> dataResult = new DataResult<GetListAccountDto>(
-				this.accountDtoMapper.accountToAccountDto(account.getData()), account.isSuccess(),
-				account.getMessage());
+		dataResult = new DataResult<GetListAccountDto>(this.accountDtoMapper.accountToAccountDto(account.getData()),
+				account.isSuccess(), account.getMessage());
 
 		return ResponseEntity.ok().lastModified(account.getData().getLastUpdateDate()).body(dataResult);
 
@@ -92,16 +110,21 @@ public class AccountsController {
 			@PathVariable(name = "senderAccountNumber") String senderAccountNumber,
 			@RequestBody @Valid CreateTransferRequest createTransferRequest) {
 
+		DataResult<GetListAccountDto> dataResult;
+
+		if (checkAuthUser(senderAccountNumber)) {
+			dataResult = new DataResult<GetListAccountDto>(null, false, "Invalid Account Number");
+			return new ResponseEntity<DataResult<GetListAccountDto>>(dataResult, HttpStatus.FORBIDDEN);
+		}
+
 		DataResult<Account> account = this.iAccountService.transferBetweenAccounts(senderAccountNumber,
 				createTransferRequest);
 
 		if (!account.isSuccess() || account == null) {
 			return new ResponseEntity<>((DataResult) account, HttpStatus.BAD_REQUEST);
 		}
-
-		DataResult<GetListAccountDto> dataResult = new DataResult<GetListAccountDto>(
-				this.accountDtoMapper.accountToAccountDto(account.getData()), account.isSuccess(),
-				account.getMessage());
+		dataResult = new DataResult<GetListAccountDto>(this.accountDtoMapper.accountToAccountDto(account.getData()),
+				account.isSuccess(), account.getMessage());
 
 		return ResponseEntity.ok().lastModified(account.getData().getLastUpdateDate()).body(dataResult);
 
@@ -110,8 +133,16 @@ public class AccountsController {
 	@DeleteMapping("/account/{accountNumber}")
 	public ResponseEntity<Result> delete(@PathVariable String accountNumber) {
 
+		Result result;
+
+		if (checkAuthUser(accountNumber)) {
+			result = new Result(false, "Invalid Account Number");
+			return new ResponseEntity<Result>(result, HttpStatus.FORBIDDEN);
+		}
 		DataResult<Account> account = this.iAccountService.delete(accountNumber);
-		Result result = new Result(account.isSuccess(), account.getMessage());
+
+		result = new Result(account.isSuccess(), account.getMessage());
+
 		if (!account.isSuccess() || account == null) {
 			return new ResponseEntity<>(result, HttpStatus.BAD_REQUEST);
 
@@ -136,12 +167,29 @@ public class AccountsController {
 	private ResponseEntity<DataResult<List<GetListLogDto>>> getLogs(
 			@PathVariable(name = "accountNumber") String accountNumber) {
 
-		DataResult<List<GetListLogDto>> dataResult = this.iAccountService.getAccountLogsByAccountNumber(accountNumber);
+		DataResult<List<GetListLogDto>> dataResult;
 
+		if (checkAuthUser(accountNumber)) {
+			dataResult = new DataResult<List<GetListLogDto>>(null, false, "Invalid Account Number");
+			return new ResponseEntity<DataResult<List<GetListLogDto>>>(dataResult, HttpStatus.FORBIDDEN);
+		}
+
+		dataResult = this.iAccountService.getAccountLogsByAccountNumber(accountNumber);
 		if (dataResult.isSuccess()) {
 			return ResponseEntity.ok().body(dataResult);
 		}
 		return new ResponseEntity<>(dataResult, HttpStatus.NOT_FOUND);
 	}
 
+	private boolean checkAuthUser(String accountNumber) {
+
+		int accountUserId = this.iAccountService.getByAccountNumber(accountNumber).getData().getUserId();
+
+		MyCustomUser authUser = (MyCustomUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+		if (authUser.getId() == accountUserId) {
+			return false;
+		}
+		return true;
+	}
 }
